scripts/generate-palettes/requirements.txt
# Pillow powers scripts/generate-palettes/run.py, which builds the palette/filter textures.
Pillow==10.4.0
